requirements.txt
appdirs==1.4.4
audioread==2.1.9
backports.weakref==1.0rc1
bidict==0.21.2
bleach==1.5.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dataclasses==0.8
decorator==4.4.2
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-SocketIO==5.0.1
future==0.18.2
gevent==20.12.1
gevent-websocket==0.10.1
google-api-core==1.24.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.34.0
h5py==2.7.1
html5lib==0.9999999
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Keras==2.0.6
libcst==0.3.16
llvmlite==0.35.0
Markdown==3.3.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numba==0.52.0
numpy==1.19.5
opencv-python==4.5.1.48
packaging==20.8
pooch==1.3.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
resampy==0.2.2
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
tensorflow==1.2.1
Theano==1.0.5
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
